[Refactor] 통합 테스트 패키지 위치 변경 #57
