+++ src/test/java/com/example/starhub/integration/PopularMeetingIntegrationTest.java
@@ -1,4 +1,4 @@
-package com.example.starhub.service;
+package com.example.starhub.integration;
 
 import com.example.starhub.dto.request.CreateMeetingRequestDto;
 import com.example.starhub.dto.request.TechStackDto;
@@ -13,6 +13,9 @@ import com.example.starhub.repository.LikeRepository;
 import com.example.starhub.repository.MeetingRepository;
 import com.example.starhub.repository.TechStackRepository;
 import com.example.starhub.repository.UserRepository;
+import com.example.starhub.service.LikeService;
+import com.example.starhub.service.MeetingService;
+import com.example.starhub.service.PopularMeetingService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -29,7 +32,7 @@ import static org.junit.jupiter.api.Assertions.*;
 @SpringBootTest
 @Transactional
 @Nested
-class PopularMeetingServiceTest {
+class PopularMeetingIntegrationTest {
 
     @Autowired
     private UserRepository userRepository;
